Player now visaully experiences hits, which intensify based on their hit points

+++ src/GameUtils.h
@@ -21,6 +21,6 @@
 		void checkPlayerHit(Player &player);
 		void setHitTime();
 	
-		static  int hitTime;
+		static  float hitTime;
 
 #endif
